% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exemple.R
\docType{package}
\name{exemple-package}
\alias{exemple}
\alias{exemple-package}
\title{Exemple}
\description{
Ceci est la description du package Exemple.
}
\details{
Les paragraphes suivants donne différents détails.

Il peut y en avoir autant que l'on veut!
}
\author{
\strong{Maintainer}: Fournier Patrick \email{pfournier@countermail.com}

}
