% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{grosgauss}
\alias{grosgauss}
\title{Plein de N(0, 1) iid!!!}
\format{
Un \code{data.frame} de 100 rangées par 10 colonnes.
}
\usage{
grosgauss
}
\description{
Jeu de donnéesdont toutes les entrées sont indépendantes et
identiquement distribuées N(0, 1). Pour votre plaisir personnel :)
}
\examples{
head(grosgauss)

shapiro.test(unlist(grosgauss))
}
\references{
Gauss, Carolo Friderico (1809) Theoria motvs corporvm
    coelestivm in sectionibvs conicis solem ambientivm
}
\seealso{
\code{\link{rnorm}}
}
\keyword{datasets}
