% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfsr.R
\name{lfsr}
\alias{lfsr}
\title{Linear Feedback Shift Registers}
\usage{
lfsr(seed = 42L, prim_poly = 3L)
}
\arguments{
\item{seed}{entier servant de germe}

\item{prim_poly}{vecteur d'entiers représentant un polynôme
    primitif sur 32 bits

\code{prim_poly} liste les puissances associées à un coefficient
non nul. 2^0 et 2^31 sont omises.}
}
\value{
Un LFSR.
}
\description{
LFSR 32 bits destiné à la génération de nombres aléatoires.
}
