% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplication.R
\name{mult2}
\alias{mult2}
\alias{mult2.numeric}
\alias{mult2.integer}
\alias{mult2.list}
\title{Multiplie par 2}
\usage{
mult2(x, ...)

\method{mult2}{numeric}(x, ...)

\method{mult2}{integer}(x, ...)

\method{mult2}{list}(x, ...)
}
\arguments{
\item{x}{objet à multiplier par 2}
}
\value{
Un objet du même type que \code{x} correspondant au produit
    de \code{x} avec 2.
}
\description{
Fonction générique permettant de multiplier un objet par 2.
}
