% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addition.R
\name{plus1}
\alias{plus1}
\title{Additionne 1}
\usage{
plus1(x)
}
\arguments{
\item{x}{vecteur de nombres.}
}
\value{
Un vecteur contenant les même nombres que \code{x} auxquels
    on a ajouté 1.
}
\description{
Ajoute 1 à un vecteur, élément par élément.
}
\details{
La fonction \code{plus1} ajoute 1 à son argument. Pour ce faire,
elle exécute une opération mathématique connue sous le nom
d'addition.

L'addition est commutative. En conséquence, ses arguments peuvent
commuter sans pour autant changer le résultat.
}
\examples{
plus1(1:5)

}
