% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mixture.R
\name{sample_mixture}
\alias{sample_mixture}
\title{Mélange entre uniforme et chi2}
\usage{
sample_mixture(n)
}
\arguments{
\item{n}{nombre d'observations}
}
\value{
Un vecteur contenant \code{n} réalisations de
la distribution mélange.
}
\description{
Fonction qui génère d'un mélange entre une uniforme et 
une chi2 à 1 degré de liberté.
}
