% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfsr.R
\name{shift}
\alias{shift}
\alias{shift.lfsr}
\title{Met à jour un générateur aléatoire.}
\usage{
shift(rng, ...)

\method{shift}{lfsr}(rng, ...)
}
\arguments{
\item{rng}{générateur aléatoire}

\item{\dots}{arguments passés aux méthodes}
}
\description{
Met à jour un générateur aléatoire.
}
